requirements.txt
aniso8601==9.0.1
blinker==1.7.0
click==8.1.7
configparser==6.0.0
Flask==3.0.1
Flask-RESTful==0.3.10
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.4
pyodbc==5.0.1
python-dotenv==1.0.1
pytz==2023.3.post1
six==1.16.0
Werkzeug==3.0.1
